moved telegram url to the public dir
