+++ public/telegramChatLink.ts
@@ -1,0 +1,1 @@
+export const TELEGRAM_URL = "https://t.me/+vPlJICrfUOM3MTVi";
