+++ src/components/homePage/after/After.tsx
@@ -1,5 +1,9 @@
 import React from "react";
 
 export default function After() {
-  return <div>After</div>;
+  return (
+    <section className="container max-w-[1920px] scroll-mt-[82px] tabxl:scroll-mt-[124px] pt-[443px] tabxl:pt-[197px] pb-14 tabxl:pb-[82px]">
+      <p>После</p>
+    </section>
+  );
 }
