+++ src/constants/constants.ts
@@ -5,6 +5,4 @@ export const menuList = [
   { title: "Отзывы", path: "reviews" },
 ];
 
-export const TELEGRAM_URL = "https://t.me/+vPlJICrfUOM3MTVi";
-
 export const CYANIDIUM_URL = "https://www.cyanidium.dev/";
